Fix jl.Long AccumulatorFactoryInfo to use LongAccumulator
The boxed java.lang.Long stream path should mirror
longAccumulatorFactoryInfo and jIntegerAccumulatorFactoryInfo, using
LongAccumulator instead of a copy-pasted IntAccumulator type.

+++ src/library/scala/collection/convert/StreamExtensions.scala
@@ -474,7 +474,7 @@ object StreamExtensions {
     }
 
     implicit val jIntegerAccumulatorFactoryInfo: AccumulatorFactoryInfo[jl.Integer, IntAccumulator] = intAccumulatorFactoryInfo.asInstanceOf[AccumulatorFactoryInfo[jl.Integer, IntAccumulator]]
-    implicit val jLongAccumulatorFactoryInfo: AccumulatorFactoryInfo[jl.Long, IntAccumulator] = longAccumulatorFactoryInfo.asInstanceOf[AccumulatorFactoryInfo[jl.Long, IntAccumulator]]
+    implicit val jLongAccumulatorFactoryInfo: AccumulatorFactoryInfo[jl.Long, LongAccumulator] = longAccumulatorFactoryInfo.asInstanceOf[AccumulatorFactoryInfo[jl.Long, LongAccumulator]]
     implicit val jDoubleAccumulatorFactoryInfo: AccumulatorFactoryInfo[jl.Double, IntAccumulator] = doubleAccumulatorFactoryInfo.asInstanceOf[AccumulatorFactoryInfo[jl.Double, IntAccumulator]]
   }
 }
